refactor(测试计划): 优化测试计划执行完成之后的回调函数
优化测试计划执行完成之后的回调函数
